Promote selected agent chat to assistant chat
